refactor(brain): give the config validator one block per module
`policy.ts` was 2204 lines with a single 1220-line validator function -
the worst maintainability score in the repository (MI -80.6, cognitive
complexity 441) and imported by 70 files, which is why the path and all
55 exports stay exactly where they were.

The seam came from the call graph rather than from line counts: each
`resolve*` export already had a distinct consumer family, so the file
splits into one module per `_brain.yaml` block, each owning its default
table, its read-side resolver and its parser together - the trio that
always changes as a unit. Above them sit the constraint vocabulary, the
key index, the block order and the disk read. Worst cognitive complexity
in the unit is now 32; the largest file is 225 lines.

Nine defects fixed while moving: a triplicated YAML-safety loop, an
array guard written out at five sites, two bound checks with the same
message, a dead `Set` branch, a dead ternary repeated in six blocks, a
default parameter no call site overrode, a boolean check split out of
the loop that handled its four siblings, validation sitting a thousand
lines from the construction it guarded, and three parameters threaded
loose through twenty call sites.

Equivalence is proved differentially, not by inspection: the HEAD
validator and the new one were run side by side over 3135 generated
configs covering every block, every key and thirty value shapes,
comparing resolved config, key insertion order, warning list and order,
thrown error name, message and field, and the full key index. Zero
mismatches. All 341 tests importing the module pass unmodified.
